save and check address SSL information

+++ model/domain/DetailDomain.go
@@ -3,8 +3,8 @@ package domain
 import "time"
 
 type DetailDomain struct {
-	ID int
-	IDDomain int
+	ID int64
+	IDDomain int64
 	IpAddress string
 	ServerName string
 	Grade string
